[refactor] Command 인터페이스 위치 변경
