+++ shell/src/main/java/com/samsung/Command.java
@@ -1,4 +1,4 @@
-package com.samsung.testshell;
+package com.samsung;
 
 public interface Command {
     void execute(String[] cmdArgs);
