+++ shell/src/test/java/com/samsung/testshell/CommandInvokerTest.java
@@ -1,5 +1,6 @@
 package com.samsung.testshell;
 
+import com.samsung.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
